refactor: Use convert_exceptions + update tests
