+++ openedx/core/djangoapps/content_libraries/collections/rest_api/v1/views.py
@@ -19,6 +19,7 @@ from openedx_events.content_authoring.signals import (
 )
 
 from openedx.core.djangoapps.content_libraries import api, permissions
+from openedx.core.djangoapps.content_libraries.utils import convert_exceptions
 from openedx.core.djangoapps.content_libraries.serializers import (
     ContentLibraryCollectionSerializer,
     ContentLibraryCollectionCreateOrUpdateSerializer,
@@ -39,11 +40,7 @@ class LibraryCollectionsView(ModelViewSet):
         """
         Verify that the collection belongs to the library and the user has the correct permissions
         """
-        try:
-            library_obj = api.require_permission_for_library_key(library_key, user, permission)
-        except api.ContentLibraryNotFound as exc:
-            raise Http404 from exc
-
+        library_obj = api.require_permission_for_library_key(library_key, user, permission)
         collection = None
         if library_obj.learning_package_id:
             collection = authoring_api.get_learning_package_collections(
@@ -51,6 +48,7 @@ class LibraryCollectionsView(ModelViewSet):
             ).filter(id=collection_id).first()
         return collection
 
+    @convert_exceptions
     def retrieve(self, request, *args, **kwargs):
         """
         Retrieve the Content Library Collection
@@ -74,6 +72,7 @@ class LibraryCollectionsView(ModelViewSet):
         serializer = self.get_serializer(collection)
         return Response(serializer.data)
 
+    @convert_exceptions
     def list(self, request, *args, **kwargs):
         """
         List Collections that belong to Content Library
@@ -85,17 +84,15 @@ class LibraryCollectionsView(ModelViewSet):
         # Check if user has permissions to view collections by checking if user
         # has permission to view the Content Library they belong to
         library_key = LibraryLocatorV2.from_string(lib_key_str)
-        try:
-            content_library = api.require_permission_for_library_key(
-                library_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY
-            )
-        except api.ContentLibraryNotFound as exc:
-            raise Http404 from exc
+        content_library = api.require_permission_for_library_key(
+            library_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY
+        )
 
         collections = authoring_api.get_learning_package_collections(content_library.learning_package.id)
         serializer = self.get_serializer(collections, many=True)
         return Response(serializer.data)
 
+    @convert_exceptions
     def create(self, request, *args, **kwargs):
         """
         Create a Collection that belongs to a Content Library
@@ -107,12 +104,9 @@ class LibraryCollectionsView(ModelViewSet):
         # Check if user has permissions to create a collection in the Content Library
         # by checking if user has permission to edit the Content Library
         library_key = LibraryLocatorV2.from_string(lib_key_str)
-        try:
-            content_library = api.require_permission_for_library_key(
-                library_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY
-            )
-        except api.ContentLibraryNotFound as exc:
-            raise Http404 from exc
+        content_library = api.require_permission_for_library_key(
+            library_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY
+        )
 
         create_serializer = ContentLibraryCollectionCreateOrUpdateSerializer(data=request.data)
         create_serializer.is_valid(raise_exception=True)
@@ -134,6 +128,7 @@ class LibraryCollectionsView(ModelViewSet):
 
         return Response(serializer.data)
 
+    @convert_exceptions
     def partial_update(self, request, *args, **kwargs):
         """
         Update a Collection that belongs to a Content Library
@@ -170,6 +165,7 @@ class LibraryCollectionsView(ModelViewSet):
 
         return Response(serializer.data)
 
+    @convert_exceptions
     def destroy(self, request, *args, **kwargs):
         """
         Deletes a Collection that belongs to a Content Library
